Move Sources/RunnerBar/Services to RunnerBarCore (#1623)
- Add Sources/RunnerBarCore/Services/TerminalLauncher.swift (public)
- Add Sources/RunnerBarCore/Services/LoginItem.swift (public)
- Add Sources/RunnerBarCore/Services/FailureHookRunnerAdapters.swift (public)
- Add Sources/RunnerBarCore/Services/FailureHookRunner.swift (public)
- Tombstone all four app-target originals

None of the files import AppKit. TerminalLauncher uses NSAppleScript
which resolves from Foundation on macOS. LoginItem imports
ServiceManagement only. FailureHookRunner and FailureHookRunnerAdapters
import Foundation + RunnerBarCore only.
